+++ FlareECS/src/FlareECS/Entity/Entity.h
@@ -42,7 +42,7 @@ namespace Flare
 	{
 		Entity Id;
 
-		size_t RegistryIndex;
+		uint32_t RegistryIndex;
 		ArchetypeId Archetype;
 		size_t BufferIndex;
 	};
